MenuButton SFX, soundtrack for menu, cutscene, endings
Introduced a new MenuButton sound effect and updated AudioManager to support it. Added a QuitButton script and hooked it up in the MainMenu scene. Also instantiated AudioManager and SoundTrackPlayer prefabs in all ending and main menu scenes. Minor adjustments to OpeningScene, including removal of unused parallax properties and color fixes.